Moving SrvKeyspace rebuild to topotools (so tabletmanager can use it).
