+++ go/vt/topotools/rebuild_shard.go
