+++ go/vt/topotools/rebuild_shard_test.go
