+++ go/vt/wrangler/rebuild.go
@@ -5,19 +5,12 @@
 package wrangler
 
 import (
-	"bytes"
-	"encoding/hex"
 	"fmt"
 	"sync"
 
-	"github.com/youtube/vitess/go/vt/concurrency"
-	"github.com/youtube/vitess/go/vt/tabletmanager/actionnode"
 	"github.com/youtube/vitess/go/vt/topo"
-	"github.com/youtube/vitess/go/vt/topo/topoproto"
 	"github.com/youtube/vitess/go/vt/topotools"
 	"golang.org/x/net/context"
-
-	pb "github.com/youtube/vitess/go/vt/proto/topodata"
 )
 
 // RebuildShardGraph rebuilds the serving and replication rollup data data while locking
@@ -27,186 +20,8 @@ func (wr *Wrangler) RebuildShardGraph(ctx context.Context, keyspace, shard strin
 }
 
 // RebuildKeyspaceGraph rebuilds the serving graph data while locking out other changes.
-// If some shards were recently read / updated, pass them in the cache so
-// we don't read them again (and possible get stale replicated data)
 func (wr *Wrangler) RebuildKeyspaceGraph(ctx context.Context, keyspace string, cells []string, rebuildSrvShards bool) error {
-	actionNode := actionnode.RebuildKeyspace()
-	lockPath, err := wr.lockKeyspace(ctx, keyspace, actionNode)
-	if err != nil {
-		return err
-	}
-
-	err = wr.rebuildKeyspace(ctx, keyspace, cells, rebuildSrvShards)
-	return wr.unlockKeyspace(ctx, keyspace, actionNode, lockPath, err)
-}
-
-// findCellsForRebuild will find all the cells in the given keyspace
-// and create an entry if the map for them
-func (wr *Wrangler) findCellsForRebuild(ki *topo.KeyspaceInfo, shardMap map[string]*topo.ShardInfo, cells []string, srvKeyspaceMap map[string]*pb.SrvKeyspace) {
-	for _, si := range shardMap {
-		for _, cell := range si.Cells {
-			if !topo.InCellList(cell, cells) {
-				continue
-			}
-			if _, ok := srvKeyspaceMap[cell]; !ok {
-				srvKeyspaceMap[cell] = &pb.SrvKeyspace{
-					ShardingColumnName: ki.ShardingColumnName,
-					ShardingColumnType: ki.ShardingColumnType,
-					ServedFrom:         ki.ComputeCellServedFrom(cell),
-					SplitShardCount:    ki.SplitShardCount,
-				}
-			}
-		}
-	}
-}
-
-// This function should only be used with an action lock on the keyspace
-// - otherwise the consistency of the serving graph data can't be
-// guaranteed.
-//
-// Take data from the global keyspace and rebuild the local serving
-// copies in each cell.
-func (wr *Wrangler) rebuildKeyspace(ctx context.Context, keyspace string, cells []string, rebuildSrvShards bool) error {
-	wr.logger.Infof("rebuildKeyspace %v", keyspace)
-
-	ki, err := wr.ts.GetKeyspace(ctx, keyspace)
-	if err != nil {
-		return err
-	}
-
-	var shardCache map[string]*topo.ShardInfo
-	if rebuildSrvShards {
-		shards, err := wr.ts.GetShardNames(ctx, keyspace)
-		if err != nil {
-			return nil
-		}
-
-		// Rebuild all shards in parallel, save the shards
-		shardCache = make(map[string]*topo.ShardInfo)
-		wg := sync.WaitGroup{}
-		mu := sync.Mutex{}
-		rec := concurrency.FirstErrorRecorder{}
-		for _, shard := range shards {
-			wg.Add(1)
-			go func(shard string) {
-				if shardInfo, err := wr.RebuildShardGraph(ctx, keyspace, shard, cells); err != nil {
-					rec.RecordError(fmt.Errorf("RebuildShardGraph failed: %v/%v %v", keyspace, shard, err))
-				} else {
-					mu.Lock()
-					shardCache[shard] = shardInfo
-					mu.Unlock()
-				}
-				wg.Done()
-			}(shard)
-		}
-		wg.Wait()
-		if rec.HasErrors() {
-			return rec.Error()
-		}
-
-	} else {
-		shardCache, err = wr.ts.FindAllShardsInKeyspace(ctx, keyspace)
-		if err != nil {
-			return err
-		}
-	}
-
-	// Build the list of cells to work on: we get the union
-	// of all the Cells of all the Shards, limited to the provided cells.
-	//
-	// srvKeyspaceMap is a map:
-	//   key: cell
-	//   value: topo.SrvKeyspace object being built
-	srvKeyspaceMap := make(map[string]*pb.SrvKeyspace)
-	wr.findCellsForRebuild(ki, shardCache, cells, srvKeyspaceMap)
-
-	// Then we add the cells from the keyspaces we might be 'ServedFrom'.
-	for _, ksf := range ki.ServedFroms {
-		servedFromShards, err := wr.ts.FindAllShardsInKeyspace(ctx, ksf.Keyspace)
-		if err != nil {
-			return err
-		}
-		wr.findCellsForRebuild(ki, servedFromShards, cells, srvKeyspaceMap)
-	}
-
-	// for each entry in the srvKeyspaceMap map, we do the following:
-	// - read the SrvShard structures for each shard / cell
-	// - if not present, build an empty one from global Shard
-	// - compute the union of the db types (replica, master, ...)
-	// - sort the shards in the list by range
-	// - check the ranges are compatible (no hole, covers everything)
-	for cell, srvKeyspace := range srvKeyspaceMap {
-		for _, si := range shardCache {
-			servedTypes := si.GetServedTypesPerCell(cell)
-
-			// for each type this shard is supposed to serve,
-			// add it to srvKeyspace.Partitions
-			for _, tabletType := range servedTypes {
-				partition := topoproto.SrvKeyspaceGetPartition(srvKeyspace, tabletType)
-				if partition == nil {
-					partition = &pb.SrvKeyspace_KeyspacePartition{
-						ServedType: tabletType,
-					}
-					srvKeyspace.Partitions = append(srvKeyspace.Partitions, partition)
-				}
-				partition.ShardReferences = append(partition.ShardReferences, &pb.ShardReference{
-					Name:     si.ShardName(),
-					KeyRange: si.KeyRange,
-				})
-			}
-		}
-
-		if err := wr.orderAndCheckPartitions(cell, srvKeyspace); err != nil {
-			return err
-		}
-	}
-
-	// and then finally save the keyspace objects
-	for cell, srvKeyspace := range srvKeyspaceMap {
-		wr.logger.Infof("updating keyspace serving graph in cell %v for %v", cell, keyspace)
-		if err := wr.ts.UpdateSrvKeyspace(ctx, cell, keyspace, srvKeyspace); err != nil {
-			return fmt.Errorf("writing serving data failed: %v", err)
-		}
-	}
-	return nil
-}
-
-// orderAndCheckPartitions will re-order the partition list, and check
-// it's correct.
-func (wr *Wrangler) orderAndCheckPartitions(cell string, srvKeyspace *pb.SrvKeyspace) error {
-
-	// now check them all
-	for _, partition := range srvKeyspace.Partitions {
-		tabletType := partition.ServedType
-		topoproto.ShardReferenceArray(partition.ShardReferences).Sort()
-
-		// check the first Start is MinKey, the last End is MaxKey,
-		// and the values in between match: End[i] == Start[i+1]
-		first := partition.ShardReferences[0]
-		if first.KeyRange != nil && len(first.KeyRange.Start) != 0 {
-			return fmt.Errorf("keyspace partition for %v in cell %v does not start with min key", tabletType, cell)
-		}
-		last := partition.ShardReferences[len(partition.ShardReferences)-1]
-		if last.KeyRange != nil && len(last.KeyRange.End) != 0 {
-			return fmt.Errorf("keyspace partition for %v in cell %v does not end with max key", tabletType, cell)
-		}
-		for i := range partition.ShardReferences[0 : len(partition.ShardReferences)-1] {
-			fn := partition.ShardReferences[i].KeyRange == nil
-			sn := partition.ShardReferences[i+1].KeyRange == nil
-			if fn != sn {
-				return fmt.Errorf("shards with unconsistent KeyRanges for %v in cell %v at shard %v", tabletType, cell, i)
-			}
-			if fn {
-				// this is the custom sharding case, all KeyRanges must be nil
-				continue
-			}
-			if bytes.Compare(partition.ShardReferences[i].KeyRange.End, partition.ShardReferences[i+1].KeyRange.Start) != 0 {
-				return fmt.Errorf("non-contiguous KeyRange values for %v in cell %v at shard %v to %v: %v != %v", tabletType, cell, i, i+1, hex.EncodeToString(partition.ShardReferences[i].KeyRange.End), hex.EncodeToString(partition.ShardReferences[i+1].KeyRange.Start))
-			}
-		}
-	}
-
-	return nil
+	return topotools.RebuildKeyspace(ctx, wr.logger, wr.ts, keyspace, cells, rebuildSrvShards)
 }
 
 func strInList(sl []string, s string) bool {
